Add additional paths
